+++ src/main/groovy/info/solidsoft/gradle/pitest/PitestTask.groovy
@@ -32,6 +32,7 @@ import org.gradle.api.tasks.JavaExec
 import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.OutputFile
+import org.gradle.api.tasks.options.Option
 
 /**
  * Gradle task implementation for Pitest.
@@ -122,11 +123,19 @@ class PitestTask extends JavaExec {
 
     @Input
     @Optional
+    final Property<Boolean> skipFailingTests
+
+    @Input
+    @Optional
     final SetProperty<String> includedGroups
 
     @Input
     @Optional
     final SetProperty<String> excludedGroups
+
+    @Input
+    @Optional
+    final SetProperty<String> includedTestMethods
 
     @InputFiles
     Set<File> sourceDirs
@@ -206,7 +215,19 @@ class PitestTask extends JavaExec {
 
     @Input
     @Optional
+    final Property<Boolean> useClasspathJar
+
+    @Input
+    @Optional
     final ListProperty<String> features
+
+    @Incubating
+    @Option(option = "additionalFeatures", description = "Additional PIT features to be appended to those placed in configuration")
+    List<String> additionalFeatures //ListProperty<String> cannot be used with @Option - https://github.com/gradle/gradle/issues/10517
+
+    @Incubating
+    @Option(option = "targetTests", description = "Tests classes to use. Overrides 'testClasses' defined in configuration")
+    List<String> overriddenTargetTests  //should be Set<String> or SetProperty but it's not supported in Gradle as of 5.6.1
 
     PitestTask() {
         ObjectFactory of = project.objects
@@ -231,8 +252,10 @@ class PitestTask extends JavaExec {
         childProcessJvmArgs = of.listProperty(String)
         outputFormats = of.setProperty(String)
         failWhenNoMutations = of.property(Boolean)
+        skipFailingTests = of.property(Boolean)
         includedGroups = of.setProperty(String)
         excludedGroups = of.setProperty(String)
+        includedTestMethods = of.setProperty(String)
 //        sourceDirs = of.fileProperty()
         detectInlinedCode = of.property(Boolean)
         timestampedReports = of.property(Boolean)
@@ -249,6 +272,7 @@ class PitestTask extends JavaExec {
 //        mutableCodePaths = of.setProperty(File)
         pluginConfiguration = of.mapProperty(String, String)
         maxSurviving = of.property(Integer)
+        useClasspathJar = of.property(Boolean)
         useAdditionalClasspathFile = of.property(Boolean)
 //        additionalClasspathFile = of.fileProperty()
         features = of.listProperty(String)
@@ -278,7 +302,7 @@ class PitestTask extends JavaExec {
 //        map['reportDir'] = reportDir.getOrNull()?.toString()
         map['reportDir'] = getReportDir().toString()
         map['targetClasses'] = targetClasses.get().join(',')
-        map['targetTests'] = optionalCollectionAsString(targetTests)
+        map['targetTests'] = overriddenTargetTests ? overriddenTargetTests.join(',') : optionalCollectionAsString(targetTests)
         map['dependencyDistance'] = optionalPropertyAsString(dependencyDistance)
         map['threads'] = optionalPropertyAsString(threads)
         map['mutateStaticInits'] = optionalPropertyAsString(mutateStaticInits)
@@ -295,8 +319,10 @@ class PitestTask extends JavaExec {
         map['jvmArgs'] = optionalCollectionAsString(childProcessJvmArgs)
         map['outputFormats'] = optionalCollectionAsString(outputFormats)
         map['failWhenNoMutations'] = optionalPropertyAsString(failWhenNoMutations)
+        map['skipFailingTests'] = optionalPropertyAsString(skipFailingTests)
         map['includedGroups'] = optionalCollectionAsString(includedGroups)
         map['excludedGroups'] = optionalCollectionAsString(excludedGroups)
+        map['includedTestMethods'] = optionalCollectionAsString(includedTestMethods)
         map['sourceDirs'] = (getSourceDirs()*.path)?.join(',')
         map['detectInlinedCode'] = optionalPropertyAsString(detectInlinedCode)
         map['timestampedReports'] = optionalPropertyAsString(timestampedReports)
@@ -308,7 +334,8 @@ class PitestTask extends JavaExec {
         map['includeLaunchClasspath'] = Boolean.FALSE.toString()   //code to analyse is passed via classPath
         map['jvmPath'] = getJvmPath()?.path
         map['maxSurviving'] = optionalPropertyAsString(maxSurviving)
-        map['features'] = optionalCollectionAsString(features)
+        map['useClasspathJar'] = optionalPropertyAsString(useClasspathJar)
+        map['features'] = (features.getOrElse([]) + (additionalFeatures?: []))?.join(',')
         map.putAll(prepareMapWithClasspathConfiguration())
         map.putAll(prepareMapWithIncrementalAnalysisConfiguration())